requirements.txt
openai
python-dotenv
qdrant-client
langchain
langchain-text-splitters
pypdf2
fastapi
uvicorn
python-multipart
pyjwt
requests
cryptography
psycopg2-binary
boto3
supabase
